LUDOC-549 sec: doc update for setting permanent nm filesets
This patch updates the "Isolating Clients to a Sub-directory Tree"
secion of the Lustre Operations Manual to update how permanent
filesets are handled in Lustre 2.17 and that the use of "lctl
set_param (-P)" for filesets is discouraged. "lctl
nodemap_set_fileset" should be used instead (as described in
LU-18357).

Signed-off-by: Marc Vef <mvef@whamcloud.com>
Change-Id: If42bf72ababe2ac9d4e643c42ca3f0cd6c130a4a
Reviewed-on: https://review.whamcloud.com/c/doc/manual/+/58367
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>