LU-13814 clio: add cl_dio_pages_init
Just like the cl_page it's replacing, the cl_dio_pages
struct needs various pieces of information from the
different layers of the cl_object in order to do the IO.

This means we need a cl_dio_pages init, analogous to
cl_page_alloc and coo_pages_init.

Note this does not implement coo_pages_init for any layers,
it just moves parts of the existing init code.

coo_dio_pages_init will be implemented in the next
patch.

Signed-off-by: Patrick Farrell <pfarrell@whamcloud.com>
Change-Id: I1fcf407b16d4077d94c7ba5afbc63bdd3fb3dfb4
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/52109
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>