LU-18876 osp: rename conflicting ping_show
The function ping_show() is defined twice: once is OSP
and another time in Portals RPC. This causes a linker
error when the Lustre modules are built into the static
kernel image. Rename the ping_show() in OSP to
osp_ping_show() and adjust the attribute definition to
match.

Test-Parameters: trivial
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: Ibe21f0714a22637f365c9bd1f5cac31608e65cbf
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58628
Reviewed-by: Alex Zhuravlev <bzzz@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>