LU-18175 lustre: back port shrinker debugfs support.
The Linux kernel has developed a shrinker API to help deal
with a system that under memory pressure. Lustre uses this
shrinker API for several subsystems. While it helps this
API lacked the ability to manually trigger a cleanup. Now
with newer kernels a debugfs interface has been developed
just for this purpose. Back port this interface so Lustre
adminstrators can use this new feature to lighten the
load related to Lustre.

The following link explains this new debugfs interface:

https://lwn.net/Articles/896274

and how to use it:

https://docs.kernel.org/admin-guide/mm/shrinker_debugfs.html

For older kernels since Lustre shrinkers are not memcg aware
we remove that code since older kernels lack newer memcg
features. Sadly memcg functionality is not exported so
even for newer kernels we can't support it.

Note this is the basic functionality present in newer
kernels. A follow on patch will provide support in our
tools for this.

Test-Parameter: trivial
Signed-off-by: James Simmons <jsimmons@infradead.org>
Change-Id: I70aeb4af15c46036e51e362a69602babfd5ea05f
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/56174
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Brian Atkinson <batkinson@lanl.gov>