LU-16350 ldiskfs: update for kernel 6.12
Update patches
   ext4-data-in-dirent.patch
   ext4-ext-merge.patch
   ext4-filename-encode.patch
   ext4-prealloc.patch
for kernel 6.12.

Linux commit v6.11-rc4-18-g4f05ee2f82b4
  ext4: store cookie in private data

If file private_data is not set and an open() handler
is defined then the file must be open()'d before proceeding
to iterate() over the directory.

This corresponds with the dropping of file->f_version member
which is replaced by the cookie in file->private_data.

Note that the file->open() handler will load the expected
cookie into file->private_data.

For older kernels the open() handler is not provided.

HPE-bug-id: LUS-11376
Test-Parameters: trivial
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Ie64bb701f45a024dce4caefe0e3ee673e6cfe25a
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/57126
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Timothy Day <timday@amazon.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>