LU-18933 gss: allow expired ctx from server for LDLM callback
When a client receives an ldlm callback request to release a lock,
the server might use an expired reverse context to send this request.
The server might not have any other choice, as it only has a reverse
context and cannot refresh it explicitly. And if the client refuses to
use the associated gss context, it fails to reply to the ldlm callback
request and gets evicted.
In order to avoid this eviction, the client needs to accept the
expired gss context associated with the server request. The client
cannot refresh its own context immediately, as it would not match the
reverse context used by the server. But the client context is going to
be refreshed right after that, along with the subsequent ldlm cancel
request.

sanity-krb5 test_201 is added to exercise this use case.

Test-Parameters: kerberos=true testlist=sanity-krb5
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I9bc82731cb7013e07cc09295cb488f52a0034ea9
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/58840
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Aurelien Degremont <adegremont@nvidia.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>