LU-18963 lnet: remove erroneous lnet_inet_enumerate() CWARN
Upon finding a DOWN interface, LNet will emit a message like
"lnet: Ignoring interface eth0: it's down". However, this is
not true - LNet doesn't ignore eth0; it chooses to spam the
kernel log.

On an instance with a large number of network interfaces,
many of them may be down. And many of them could be designated
for Lustre. As a result, this error could be printed hundreds
of times. But the error doesn't provide actionable advice to
the user; they aren't attempting to set up those interfaces and
hence don't care about them.

Downgrade this warning to a CDEBUG. lnet_inet_select() will emit
an appropriate warning, if needed.

Test-Parameters: trivial
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: I17942c2eeb37ffa5c6b7d9c107e9eae108105115
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59042
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>