LU-18964 tests: performance-sanity test_2 fixed large
test_2 in performance-sanity previously had mdsrate-create-large
meaning that the files themselves were large, however for performance
testing this doesn't make much sense as this is largely dependant on
the system IO. Instead large now means a large number of files (1M),
which was the original goal of these tests.

The SLOW/zfs check was also moved from test_1 to test_2 as it had
been left when the tests were renumbered.

Fixes: 01d16dadab ("LU-14697 tests: change performance-sanity to use mdtest")
Test-Parameters: trivial testlist=performance-sanity env=ONLY=1+2
Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Change-Id: I602b76c050d29b5d401971123caa73e5d4844342
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59047
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>