LU-18238 llite: deregister HSM agent on close()
This patch automatically unregisters the HSM agent when closing the
dir FD mountpoint (opened by llapi_hsm_copytool_register()).
So it would unregister the agent if the copytool has crashed and
was unable to call llapi_hsm_copytool_unregister().

This should avoid the MDTs to send HSM requests to an already crashed
copytool.

Notes: the MDT unregisters an agent if it returns -EPIPE to an
HSM request (the Lustre client has lost the connection with the
copytool). But this causes the HSM request (and the HSM actions)
to fail.

Add regression test sanity-hsm 106b.

Test-Parameters: testlist=sanity-hsm env=ONLY=106a,ONLY_REPEAT=20
Test-Parameters: testlist=sanity-hsm env=ONLY=106b,ONLY_REPEAT=20
Signed-off-by: Etienne AUJAMES <eaujames@ddn.com>
Change-Id: I9d81d758e423960dcb00e7f45d7c8c30651279a7
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59067
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>