LU-18976 gss: detect dns error in lsvcgssd
Use proper NI_NAMEREQD flag with getnameinfo() in lsvcgssd when
resolving IP addresses, for it to return an error if DNS resolution fails.

That way, the error is propagated and Lustre knows a DNS failure happens
and can report it better. If not, it fails latter and reports a very unclear
message.

Test-Parameters: kerberos=true testlist=sanity-krb5
Change-Id: Iaa9e718c056ea742d8695048e43bdeeb3205f0dd
Signed-off-by: Aurelien Degremont <adegremont@nvidia.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59095
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Bruno Faccini <bfaccini@nvidia.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>