LU-16946 tests: test_56oe sanity timing issues
sanity.sh test_56oe seems to be failing when the operations between
finding the time boundaries and retreiving the files created within
that amount of time exceeds around 2 seconds.

To fix the timing issues the tests is tried multiple times and if it
continues to fail then it will be skipped. The debug message has also
been updated to provide more useful information if this test
continues to fail in the future.

Test-Paramters: trivial, testlist=sanity env=ONLY=56oe

Signed-off-by: Frederick Dilger <fdilger@whamcloud.com>
Fixes: 28bad5e2a9 ("LU-16946 utils: allow lfs find time within a range")
Change-Id: I4ff52c255ccb23eb112d8a6e874f54c8f1694f98
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59124
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>