LU-18958 osd: improve quota declarations in setattr
Change osd_declare_attr_set() to avoid declaring transaction
credits for quota IDs that are not changing.  It is still
necessary to reserve transaction credits even if bspace=0,
because the inode accounting is still changed in this case.

This also quiets a bogus static analysis warning about 'bspace'
being uninitialized, even though the condition to enter both
code blocks is the same.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iacc11b41abfe77691639f3e901fc6a6f88505e10
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59186
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Sergey Cheremencev <scherementsev@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>