LU-19037 osc: handle osc_page not removed from radix tree
Initialize the radix tree of osc_object::oo_tree.

There are rare cases that osc_page is not removed from the radix tree
while its corresponding vmpage is release from cache, and when a page
for the same index is needed, a new osc_page is created,
it would find there is an existing item in the radix tree and -EEXIST
is returned and the IO would loop this page allocation again and
again.

This patch would explicitly call cl_page_delete() in page discard,
lest generic_error_remove_page() failed to reach folio_invalidate()
so that cl_page_delete() would not be called.

Add error messages when the radix tree insertion/deletion of osc_page
fails.

Signed-off-by: Bobi Jam <bobijam@whamcloud.com>
Change-Id: Ieb781d2f908d91aaf454df48b24fdd410cfeb348
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59660
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Qian Yingjin <qian@ddn.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Patrick Farrell <pfarrell@whamcloud.com>