LU-19079 utils: function to get nodemap name from sha256
Introduce function in gss-related utils to get the name of a nodemap
from its sha256. This is done by sending an LCFG_NODEMAP_LOOKUP_SHA
ioctl, which returns -EPERM if the nodemap does not have the
gssonly_identification property set.
Issuing the ioctl requires to find out the ioctl device to use for the
server on which the lsvcgssd daemon is running.
Rename ioc_dev (un)register functions to llapi_(un)register_ioc_*() to
avoid namespace pollution.

Test-Parameters: trivial
Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I2848d4e2e9fda00e8925ce6931cf2bd1daa5134d
Was-Change-Id: If374a2f58aff476f4014d0a5c1b40a3f9eebd475
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/61739
Reviewed-by: Andreas Dilger <adilger@thelustrecollective.com>
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>